cpufreq: remove max frequency capping from scale_freq_capacity()
After the revert of "cpufreq: Max freq invariant scheduler load-tracking
and cpu capacity support" scale_freq_capacity() is now only the setter
of the Frequency Invariant Engine (FIE). That's why the call to
scale_freq_capacity() in cpufreq_set_policy() can be deleted because
this was driving max frequency capping.

The patch changes the parameter list of scale_freq_capacity() to

 (1) set of cpus involved (cpumask)
 (2) current frequency
 (3) max possible frequency (cpuinfo related)

to not be forced to expose cpufreq internal data structures like struct
cpufreq_policy or struct cpufreq_freqs.

Change-Id: I4a8d319d2d01854409f55ab058cb5a15ed9faf9a
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>