sched: EAS: update trg_cpu to backup_cpu if no energy saving for target_cpu
If no energy saving for target_cpu in the calculation of energy_diff(),
backup_cpu will be set as the new dst_cpu for the next calculation. At this
point, we also need update the new trg_cpu as backup_cpu to make sure the
subsequent calculation of energy_diff() is correct.

Change-Id: I007891891446c78fbcb4c02346b4f4b74775ea4a
Signed-off-by: Ke Wang <ke.wang@spreadtrum.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>