rcu: Check the range of jiffies_till_xxx_fqs on setting them
Currently, the range of jiffies_till_first_fqs and jiffies_till_next_fqs
are always checked and adjusted in the loop of rcu_gp_kthread on runtime.
However, it would be better and enough to check them only on setting
them, so remove unnecessary checking and adjusting in the loop.

Additionally, add adjusting jiffies_till_first_fqs so guaranteed to be
greater than 0, which hasn't been done before.

Change-Id: I5129ffc72249dac11fa14e6df1c49b64c08988f8
Signed-off-by: Byungchul Park <byungchul.park@lge.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>